completed notifications

+++ Models/AbsenceType.cs
@@ -1,11 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Panis.Models
 {
     public class AbsenceType
     {
         [Key]
-        public int AbsenceTypeID { get; set; }
+       
+        public byte AbsenceTypeID { get; set; }
         public string Name { get; set; }
     }
 }
